Removed some of the logic functions and placed them in their own class
