+++ src/main/java/ChessLogic.java
@@ -1,0 +1,85 @@
+import org.jetbrains.annotations.NotNull;
+
+public class ChessLogic {
+    private Chessboard board;
+
+    public ChessLogic(Chessboard board) {
+        this.board = board;
+    }
+
+
+    /**
+     * Calculates if moving a piece to a position would put that teams king in check.
+     * This assumes the piece moving to the new position is a valid move.
+     * @param pieceToCheck the piece being moved.
+     * @return true if in check, false if not
+     */
+    public boolean isMoveUnsafe(int newX, int newY, @NotNull Piece pieceToCheck){
+        Piece lastPiece = board.getLastPieceMoved();
+        Move move = new Move(newX, newY, pieceToCheck, lastPiece, board);
+        boolean isMoveUnsafe = isKingInCheck(pieceToCheck.getColour());
+        move.undo();
+        return isMoveUnsafe;
+    }
+
+    public boolean isKingInCheck(@NotNull PieceColour kingToCheck){
+        King king = board.getKing(kingToCheck);
+        Coordinate kingPos = new Coordinate(king);
+        Iterable<Piece> enemyPieces = board.getAllColourPieces(PieceColour.getOtherColour(kingToCheck));
+        for(Piece piece : enemyPieces){
+            if(piece.getPossibleMoves().contains(kingPos)){
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public boolean isCheckmate(){
+        if(!isKingInCheck(board.getCurrentTurn()))
+            return false;
+        Iterable<Piece> enemyPieces = board.getAllColourPieces(board.getCurrentTurn());
+        for (Piece enemyPiece : enemyPieces) {
+            for (Coordinate move : enemyPiece.getPossibleMoves( )) {
+                if (!isMoveUnsafe(move.x(), move.y(), enemyPiece))
+                    return false;
+            }
+        }
+        return true;
+    }
+
+    public boolean isDraw(){
+        return isStalemate() ||
+                isDraw50Move() ||
+                isRepetition();
+    }
+
+    private boolean isStalemate(){
+        if(isKingInCheck(board.getCurrentTurn()))
+            return false;
+        Iterable<Piece> pieces = board.getAllColourPieces(board.getCurrentTurn());
+        for(Piece piece : pieces){
+            if(!piece.getPossibleMoves().isEmpty())
+                return false;
+        }
+        return true;
+    }
+
+    private boolean isRepetition(){
+        if(board.getNumFullMoves() < 4)
+            return false;
+        int boardState = board.getState();
+        for(int i = 0; i < 2; i++){
+            board.undoMultipleMoves(4);
+            if(boardState != board.getState()) {
+                board.redoAllMoves();
+                return false;
+            }
+        }
+        board.undoMultipleMoves(4);
+        return true;
+    }
+
+    private boolean isDraw50Move(){
+        return board.getNumHalfMoves() == 50;
+    }
+}
